DEV-1940: Render Walkontable directly to DOM, remove diffing renderer
Make the direct-DOM renderer (a fixed, viewport-sized set of nodes reused
in place) the only rendering path. Remove the ViewDiffer strategy, the
renderer-adapter factory, and the coordinate-keyed NodesPool cache, which
retained a DOM node per source coordinate ever rendered and grew memory
without bound while scrolling. Reduce NodesPool to a thin element factory
and fix multi-level row-header ordering.

Memory is now O(viewport) instead of O(rows x cols); render speed is
equal-or-faster across scenarios up to 1M rows.
