+++ handsontable/src/3rdparty/walkontable/src/utils/orderView/constants.ts
@@ -21,40 +21,3 @@ export const WORKING_SPACE_TOP = 1;
  * @type {number}
  */
 export const WORKING_SPACE_BOTTOM = 2;
-
-/**
- * Command type: no action needed, element is already in correct position.
- *
- * @type {number}
- */
-export const CMD_NONE = 0;
-/**
- * Command type: remove the element from the root node.
- *
- * @type {number}
- */
-export const CMD_REMOVE = 1;
-/**
- * Command type: append the element at the end of the root node.
- *
- * @type {number}
- */
-export const CMD_APPEND = 2;
-/**
- * Command type: prepend the element at the beginning of the root node.
- *
- * @type {number}
- */
-export const CMD_PREPEND = 3;
-/**
- * Command type: insert the element before another element in the root node.
- *
- * @type {number}
- */
-export const CMD_INSERT_BEFORE = 4;
-/**
- * Command type: replace an existing element with a new one in the root node.
- *
- * @type {number}
- */
-export const CMD_REPLACE = 5;
