--- handsontable/src/3rdparty/walkontable/src/utils/orderView/rendererAdapter/index.ts
@@ -1,19 +1,0 @@
-import type { OrderView } from '../view';
-import { isFirefox } from '../../../../../../helpers/browser';
-import { DifferBasedRendererAdapter } from './differBasedRendererAdapter';
-import { DirectDomRendererAdapter } from './directDomRendererAdapter';
-
-/**
- * Factory function to create the appropriate renderer adapter based on the browser.
- *
- * @param {OrderView} orderView The OrderView instance.
- * @returns {DifferBasedRendererAdapter|DirectDomRendererAdapter}
- */
-export function createRendererAdapter(orderView: OrderView) {
-  if (isFirefox()) {
-    return new DirectDomRendererAdapter(orderView);
-  }
-
-  return new DifferBasedRendererAdapter(orderView);
-}
-
