+++ handsontable/src/3rdparty/walkontable/src/utils/orderView/view.ts
@@ -1,12 +1,17 @@
 import { ViewSizeSet } from './viewSizeSet';
-import { ViewDiffer } from './viewDiffer';
-import { createRendererAdapter } from './rendererAdapter';
+import { WORKING_SPACE_TOP, WORKING_SPACE_BOTTOM } from './constants';
 
 /**
  * Executive model for each table renderer. It's responsible for injecting DOM nodes in a
  * specified order and adjusting the number of elements in the root node.
  *
- * Only this class have rights to juggling DOM elements within the root node (see render method).
+ * Rendering writes directly to the DOM: the root node keeps exactly `viewSize` children, and on
+ * each render cycle the existing children are reused in place (their content is overwritten by the
+ * cell renderer). There is no node pool keyed by source coordinate and no diffing strategy — those
+ * were removed because keeping a fixed, viewport-sized set of nodes is simpler and bounds memory to
+ * the viewport regardless of dataset size.
+ *
+ * Only this class has rights to juggle DOM elements within the root node (see render method).
  *
  * @class {OrderView}
  */
@@ -36,24 +41,18 @@ export class OrderView {
    */
   collectedNodes: HTMLElement[] = [];
   /**
-   * The differ which calculates the differences between current and next view. It generates
-   * commands that are processed by the OrderView (see `applyCommand` method).
-   *
-   * @type {ViewDiffer}
-   */
-  viewDiffer = new ViewDiffer(this.sizeSet);
-  /**
-   * Node type which the order view will manage while rendering the DOM elements.
+   * Node type which the order view will manage while rendering the DOM elements (uppercased to
+   * match `Element.tagName`).
    *
    * @type {string}
    */
   childNodeType;
   /**
-   * The renderer adapter that handles browser-specific rendering logic.
+   * The visual index of the currently processed child within a render cycle.
    *
-   * @type {StandardRendererAdapter|FirefoxRendererAdapter}
+   * @type {number}
    */
-  rendererAdapter;
+  visualIndex = 0;
 
   /**
    * Creates a new OrderView instance.
@@ -65,8 +64,7 @@ export class OrderView {
   constructor(rootNode: HTMLElement, nodesPool: (index?: number) => HTMLElement, childNodeType: string) {
     this.rootNode = rootNode;
     this.nodesPool = nodesPool;
-    this.childNodeType = childNodeType;
-    this.rendererAdapter = createRendererAdapter(this);
+    this.childNodeType = typeof childNodeType === 'string' ? childNodeType.toUpperCase() : childNodeType;
   }
 
   /**
@@ -128,26 +126,102 @@ export class OrderView {
   }
 
   /**
-   * Setups and prepares all necessary properties and start the rendering process.
+   * Returns the count of children in the root node that belong to this view (for shared root nodes
+   * only the children matching this view's node type and working space are counted).
+   *
+   * @returns {number}
+   */
+  getRenderedChildCount() {
+    const { rootNode, sizeSet } = this;
+    let childElementCount = 0;
+
+    if (this.isSharedViewSet()) {
+      let element = rootNode.firstElementChild;
+
+      while (element) {
+        if (element.tagName === this.childNodeType) {
+          childElementCount += 1;
+
+        } else if (sizeSet.isPlaceOn(WORKING_SPACE_TOP)) {
+          break;
+        }
+        element = element.nextElementSibling;
+      }
+    } else {
+      childElementCount = rootNode.childElementCount;
+    }
+
+    return childElementCount;
+  }
+
+  /**
+   * Setups and prepares all necessary properties and starts the rendering process. It adjusts the
+   * number of children in the root node to the next view size, creating or removing nodes as needed.
    * This method has to be called only once (at the start) for the render cycle.
    */
   start() {
-    this.rendererAdapter.start();
+    this.collectedNodes.length = 0;
+    this.visualIndex = 0;
+
+    const { rootNode, sizeSet } = this;
+    const isShared = this.isSharedViewSet();
+    const { nextSize } = sizeSet.getViewSize();
+
+    let childElementCount = this.getRenderedChildCount();
+
+    while (childElementCount < nextSize) {
+      const newNode = this.nodesPool();
+
+      if (!isShared || sizeSet.isPlaceOn(WORKING_SPACE_BOTTOM)) {
+        rootNode.appendChild(newNode);
+      } else {
+        rootNode.insertBefore(newNode, rootNode.firstChild);
+      }
+      childElementCount += 1;
+    }
+
+    const isSharedPlacedOnTop = isShared && sizeSet.isPlaceOn(WORKING_SPACE_TOP);
+
+    while (childElementCount > nextSize) {
+      const childToRemove = isSharedPlacedOnTop ? rootNode.firstChild : rootNode.lastChild;
+
+      if (childToRemove) {
+        rootNode.removeChild(childToRemove);
+      }
+      childElementCount -= 1;
+    }
   }
 
   /**
-   * Renders the DOM element based on visual index (which is calculated internally).
+   * Renders the DOM element for the current visual index by reusing the child already in place.
+   * If a child of the wrong node type occupies the slot, it is replaced with a freshly created one.
    * This method has to be called as many times as the size count is met (to cover all previously rendered DOM elements).
    */
   render() {
-    this.rendererAdapter.render();
+    const { rootNode, sizeSet } = this;
+    let visualIndex = this.visualIndex;
+
+    if (this.isSharedViewSet() && sizeSet.isPlaceOn(WORKING_SPACE_BOTTOM) && sizeSet.sharedSize) {
+      visualIndex += sizeSet.sharedSize.nextSize;
+    }
+
+    let node = rootNode.childNodes[visualIndex] as HTMLElement;
+
+    if (node.tagName !== this.childNodeType) {
+      const newNode = this.nodesPool();
+
+      node.replaceWith(newNode);
+      node = newNode;
+    }
+
+    this.collectedNodes.push(node);
+    this.visualIndex += 1;
   }
 
   /**
-   * Ends the render process.
-   * This method has to be called only once (at the end) for the render cycle.
+   * Ends the render process. No cleanup is required — all work is done in `start()` and `render()`.
    */
   end() {
-    this.rendererAdapter.end();
+    // intentionally empty
   }
 }
