+++ handsontable/src/3rdparty/walkontable/test/spec/utils/orderView/view.spec.js
@@ -5,7 +5,7 @@ describe('Walkontable.OrderView', () => {
 
     nodesPool.setRootDocument(document);
 
-    const orderView = new Walkontable.OrderView(rootNode, sourceIndex => nodesPool.obtain(sourceIndex));
+    const orderView = new Walkontable.OrderView(rootNode, sourceIndex => nodesPool.obtain(sourceIndex), childNodeType);
 
     return { orderView, rootNode };
   }
