+++ handsontable/src/3rdparty/walkontable/test/unit/utils/orderView/sharedView.unit.ts
@@ -1,7 +1,5 @@
 import { SharedOrderView } from 'walkontable/utils/orderView/sharedView';
 import { ViewSizeSet } from 'walkontable/utils/orderView/viewSizeSet';
-import { ViewDiffer } from 'walkontable/utils/orderView/viewDiffer';
-import { DifferBasedRendererAdapter } from 'walkontable/utils/orderView/rendererAdapter/differBasedRendererAdapter';
 
 /**
  *
@@ -22,9 +20,7 @@ describe('SharedOrderView', () => {
     expect(orderView.nodesPool).toBe(nodeFactoryFunction);
     expect(orderView.sizeSet).toBeInstanceOf(ViewSizeSet);
     expect(orderView.collectedNodes).toEqual([]);
-    expect(orderView.viewDiffer).toBeInstanceOf(ViewDiffer);
-    expect(orderView.rendererAdapter).toBeInstanceOf(DifferBasedRendererAdapter);
-    expect(orderView.rendererAdapter.leads).toEqual([]);
+    expect(orderView.visualIndex).toBe(0);
   });
 
   it('should correctly prepend another OrderView to this instance', () => {
